authentification: Move TokenAttributes to Auth services
